Create yoda.cpp
A problem that was in the Fanshawe IT contest during the 2024 winter term. Assuming two numbers had the same number of digits, the digits were compared and the lesser one was replaced with an 'X', and if at the end of the comparison, the entire number was replaced with 'X's, the number would be changed to the word 'YODA'.